Factor event types into separate file
